+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Thu Jan 15 19:08:18 PST 2015"
+buildTime = "Mon Jan 19 19:58:56 PST 2015"